+++ src/renderer/components/ContextMenu.tsx
@@ -141,6 +141,7 @@ export function ContextMenuLayer({
       ref={layerRef}
       className={`dc-context-menu-layer ${active ? 'active' : ''}`}
       onClick={cancel}
+      onContextMenuCapture={cancel}
     >
       {active && currentItems.length > 0 && (
         <ContextMenu
@@ -288,27 +289,12 @@ export function ContextMenu(props: {
       }
     }
 
-    const onOutsideClick = (ev: MouseEvent | TouchEvent) => {
-      const target = ev.target as HTMLElement
-      const isOnMenu = menuLevelEls.current.find(
-        menuEl => target === menuEl || target.parentElement === menuEl
-      )
-
-      if (!isOnMenu) {
-        closeCallback()
-      }
-    }
-
     const onResize = (_ev: UIEvent) => closeCallback()
 
     document.addEventListener('keydown', onKeyDown)
-    document.addEventListener('mousedown', onOutsideClick)
-    document.addEventListener('touchstart', onOutsideClick)
     window.addEventListener('resize', onResize)
     return () => {
       document.removeEventListener('keydown', onKeyDown)
-      document.removeEventListener('mousedown', onOutsideClick)
-      document.removeEventListener('touchstart', onOutsideClick)
       window.removeEventListener('resize', onResize)
     }
   }, [openSublevels, menuLevelEls, closeCallback])