+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - webxdc stay open when switching accounts (when you have sync all enabled, which is the default) #3621
 - Add more versions to about dialog and update the order of information #3677
 - Refactor message meta data component #3678
+- Use overlay in context menu #3682
 
 ### Fixed
 - Silently fail when notifications are not supported by OS #3613